requirements.txt
django-heroku
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
django-cors-headers==3.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
twilio==6.50.1
django-safedelete==1.0.0
django-storages==1.11.1
celery==5.1.2
zappa==0.53.0
pytest==6.2.4
pytest-django==4.4.0
drf-extensions==0.7.0
django-safedelete==1.0.0
PyJWT==2.1.0
